moves setSystemRoleInChatHistory to utils file
